split auth middleware tests accordingly
